+++ test/linalg/qr.jl
@@ -150,11 +150,11 @@ end
 
 @testset "transpose errors" begin
     @test_throws ErrorException transpose(qrfact(randn(3,3)))
-    @test_throws ErrorException ctranspose(qrfact(randn(3,3)))
+    @test_throws ErrorException adjoint(qrfact(randn(3,3)))
     @test_throws ErrorException transpose(qrfact(randn(3,3), Val(false)))
-    @test_throws ErrorException ctranspose(qrfact(randn(3,3), Val(false)))
+    @test_throws ErrorException adjoint(qrfact(randn(3,3), Val(false)))
     @test_throws ErrorException transpose(qrfact(big.(randn(3,3))))
-    @test_throws ErrorException ctranspose(qrfact(big.(randn(3,3))))
+    @test_throws ErrorException adjoint(qrfact(big.(randn(3,3))))
 end
 
 @testset "Issue 7304" begin